refactor(web): extract APP_DEV_HMR_CLIENT runtime + inject helpers to @zts/web
#2539 epic 의 PR 4/11. 기존 zts.mjs 의 브라우저 inject 코드 (~290줄) 와 HTML
inject helper (~50줄) 를 @zts/web 패키지로 이전. zts CLI 는 dev/preview/build
app 모드에서만 @zts/web 을 lazy import.

신설:
- packages/web/runtime/dev-overlay-client.mjs (~290줄): zts.mjs L814-1106 의
  APP_DEV_HMR_CLIENT 를 byte-for-byte 이전. brower-side overlay UI + sourcemap
  decoder + runtime error capture 의 string 출력. @zts/server 의 protocol 상수
  (APP_DEV_HMR_WS_PATH, HMR_MSG) 를 import 해 template literal 에 평가 — single
  source of truth.
- packages/web/runtime/dev-overlay-client.d.mts: type declaration.
- packages/web/src/inject.ts (~80줄): joinUrl, injectIntoDevHtml,
  injectAppDevHmrClient, injectAppDevBundleCssLinks, injectAppDevPipelineCssLinks.
- packages/web/src/inject.test.ts (~200줄): 21 케이스 — head/script 분기, 중복
  방지 (큰따옴표/작은따옴표), base path, ENOENT silent skip, EISDIR propagate.

수정:
- packages/core/bin/zts.mjs:
  * L740-812 의 inject 4 함수 (~73줄) 제거
  * L814-1106 의 APP_DEV_HMR_CLIENT 큰 string (~293줄) 제거
  * loadWebModule() lazy import wrapper 추가 — 미설치 시 안내 + exit 1
  * runAppDev / createAppDevController / runServe 에서 lazy import 후 사용
- packages/server/{tsconfig,package}.json — server 빌드를 zts bundle 에서 tsc 만
  으로 fallback (#2576 ZTS bundler 의 export * re-export emit 누락 회피).
- packages/server/package.json — \"private\": true 제거 (#2575 ZTS bundler 의
  workspace dep auto-inline 미지원 회피, npm publish 가능 패키지로). README 에
  internal 의도 명시.
- packages/web/package.json — server external 명시. devDeps 에 @types/node.
- root package.json — devDeps 에 @zts/web (workspace 환경에서 zts CLI 가 web
  resolve 가능하도록 root node_modules 에 link).

발견된 ZTS bundler 버그 2건 (별도 이슈 등록):
- #2575 workspace dep (workspace:*) auto-inline 못 함
- #2576 export * re-export 의 ESM emit 누락

검증:
- packages/web/src/inject.test.ts 21 케이스 통과
- packages/core/bin/zts.test.ts 의 dev/build/preview app builder 테스트 통과
  (watch/dev fail 은 main 도 동일 — 본 PR 회귀 아님, 별도 이슈)
- packages/server/src 테스트 67 케이스 (PR #3) 그대로 통과
- @zts/web 빌드 산출물에 dev-overlay-client 의 290줄 inline 확인

Refs #2539 (4/11)
