+++ package.json
@@ -32,6 +32,7 @@
     "@emotion/react": "^11.14.0",
     "@tailwindcss/postcss": "^4.2.2",
     "@tanstack/react-query": "^5",
+    "@zts/web": "workspace:*",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
     "hermes-engine-cli": "^0.12.0",
