+++ packages/server/tsconfig.json
@@ -3,7 +3,8 @@
   "compilerOptions": {
     "rootDir": "src",
     "outDir": "dist",
-    "declarationDir": "dist"
+    "declarationDir": "dist",
+    "emitDeclarationOnly": false
   },
   "include": ["src/**/*"],
   "exclude": ["src/**/*.test.ts"]
