+++ packages/web/runtime/dev-overlay-client.d.mts
@@ -1,0 +1,2 @@
+/** Browser inject 텍스트 — `<script>` body 로 사용. */
+export const APP_DEV_HMR_CLIENT: string;
